% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Igniter.R
\name{Igniter}
\alias{Igniter}
\title{Igniter}
\description{
An R6 class that performs optimization steps and predictions with a given network, loss function, target and optimizer.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Igniter-new}{\code{Igniter$new()}}
\item \href{#method-Igniter-train}{\code{Igniter$train()}}
\item \href{#method-Igniter-eval}{\code{Igniter$eval()}}
\item \href{#method-Igniter-opt_step}{\code{Igniter$opt_step()}}
\item \href{#method-Igniter-predict_step}{\code{Igniter$predict_step()}}
\item \href{#method-Igniter-clone}{\code{Igniter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Igniter-new"></a>}}
\if{latex}{\out{\hypertarget{method-Igniter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Igniter$new(network, loss_fn, target, optimizer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Igniter-train"></a>}}
\if{latex}{\out{\hypertarget{method-Igniter-train}{}}}
\subsection{Method \code{train()}}{
Set the network to training mode.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Igniter$train()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The object itself, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Igniter-eval"></a>}}
\if{latex}{\out{\hypertarget{method-Igniter-eval}{}}}
\subsection{Method \code{eval()}}{
Set the network to evaluation mode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Igniter$eval()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The object itself, invisibly
maybe this also also make this configurable.
Perform an optimization step

Performs a forward pass through the network, calculates the loss,
backpropagates gradients and updates model parameters using the optimizer.

Note that you need to ensure that the network is in \code{"train"}-mode beforehand.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Igniter-opt_step"></a>}}
\if{latex}{\out{\hypertarget{method-Igniter-opt_step}{}}}
\subsection{Method \code{opt_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Igniter$opt_step(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{\code{list(torch_tensor)}
Inputs to the network.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{list()} containing the loss value and network output as \code{torch_tensor} objects
Perform a prediction step

Performs a forward pass through the network, but does not calculate the loss.

Note that you need to ensure that the network is in \code{"eval"}-mode beforehand.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Igniter-predict_step"></a>}}
\if{latex}{\out{\hypertarget{method-Igniter-predict_step}{}}}
\subsection{Method \code{predict_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Igniter$predict_step()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{torch_tensor} containing the prediction.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Igniter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Igniter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Igniter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
